:recycle: refactor: Memedetail -> MemeInfo 폴더 이름 변경 [YW-144]
