+++ src/components/meme/MemeInfo/index.ts
@@ -1,3 +1,4 @@
 export * from "./MemeCTAList";
 export * from "./MemeDetail";
 export * from "./MemeTagList";
+export * from "./RelativeMemeList";
